refactor(agents): extract share modal utilities
